account_controler/requirements.txt
Telethon
PySocks